requirements.txt
alembic==1.7.7
amqp==5.1.1
attrs==21.2.0
cached-property==1.5.2
classic-app-layer==0.0.1
classic-aspects==0.0.1
classic-components==0.0.1
classic-http-api==0.0.2
classic-messaging==0.0.2
classic-messaging-kombu==0.0.3
classic-sql-storage==0.0.1
falcon==3.0.1
greenlet==1.1.3
gunicorn==20.1.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
kombu==5.1.0
Mako==1.2.3
MarkupSafe==2.1.1
pydantic==1.8.2
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.41
typing_extensions==4.3.0
vine==5.0.0
zipp==3.8.1
